fix: 삭제할 Notification이 없는 경우에 Post와 Comment를 삭제할 때의 오류 수정

+++ backend/sokdak/src/main/java/com/wooteco/sokdak/notification/service/NotificationService.java
@@ -70,12 +70,16 @@ public class NotificationService {
 
     public void deleteCommentNotification(Long commentId) {
         List<Long> ids = notificationRepository.findIdsByCommentId(commentId);
-        notificationRepository.deleteAllById(ids);
+        if (!ids.isEmpty()) {
+            notificationRepository.deleteAllById(ids);
+        }
     }
 
     public void deletePostNotification(Long postId) {
         List<Long> ids = notificationRepository.findIdsByPostId(postId);
-        notificationRepository.deleteAllById(ids);
+        if (!ids.isEmpty()) {
+            notificationRepository.deleteAllById(ids);
+        }
     }
 
     @Transactional(readOnly = true)
